The Text, Combination, Chat Commit
Added the ability to add to the UIAnnouncer's log. It's the start of a
chat system.

Homoginized the Player and Enemy classes even more to utilize the Unit
class better. Added a 'BaseXX' and 'GrowthXX' variable to handle
leveling up better. Made the level and experience variables much more
strongly coupled together. You shouldn't update one without updating the
other.

Allowed enemies to send an item to the chat upon spawning in. 'm_Taunt'.

Fixed the skill button not displaying that you do not have enough mana
to cast upon initialization.

Moved AIController to the proper folder.

Added the ability to control the player's rotation when standing still
using the right stick on a controller. Only work's for XInput currently.
