+++ Assets/Prefabs/Units/Goblin_Prefab.prefab
@@ -29,6 +29,7 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -4.45, y: 0.5, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
@@ -76,17 +77,22 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e4474093f4b84304d986bebbd4755256, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  m_Nameplate: {fileID: 11409874, guid: bdd14ad98f93f3a449cd18a355cc6dad, type: 2}
   m_ControllerType: 3
   m_Skills: []
   m_UnitName: Goblin
   m_UnitNickname: Goblin
   m_Faction: Enemies
-  m_MaxMana: 2
-  m_MaxDefense: 4
-  m_MaxHealth: 10
-  m_Experience: 5
-  m_Speed: 5
-  m_Nameplate: {fileID: 11409874, guid: bdd14ad98f93f3a449cd18a355cc6dad, type: 2}
+  m_BaseHealth: 5
+  m_HealthGrowth: 2.5
+  m_BaseMana: 2
+  m_ManaGrowth: 1
+  m_BaseDefense: 2
+  m_DefenseGrowth: 1
+  m_BaseLevel: 1
+  m_BaseSpeed: 5
+  m_SpeedGrowth: 0.2
+  m_Taunt: That smell...fear!
 --- !u!135 &13544926
 SphereCollider:
   m_ObjectHideFlags: 1
