+++ Assets/Prefabs/Units/Player1_Prefab.prefab
@@ -42,9 +42,10 @@ Transform:
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 154216}
-  m_LocalRotation: {x: 0, y: 0.7071066, z: 0, w: -0.70710695}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: -0, y: 0.5, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_Children:
   - {fileID: 497186}
   m_Father: {fileID: 0}
@@ -58,6 +59,7 @@ Transform:
   m_LocalRotation: {x: 0, y: 0, z: 0.42261833, w: 0.9063078}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_Children: []
   m_Father: {fileID: 480986}
   m_RootOrder: 0
@@ -120,6 +122,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f944c61a53426674db03c701eb3babb3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  m_Nameplate: {fileID: 11465284, guid: 5846026beed07984eba03daa6cd87c58, type: 2}
   m_ControllerType: 0
   m_Skills:
   - m_SkillPrefab: {fileID: 167634, guid: e4edc7fe0341b644c835f75acf8dad64, type: 2}
@@ -132,17 +135,18 @@ MonoBehaviour:
     m_CooldownReduction: 0
     m_RemainingCooldown: 0
   m_UnitName: Player1
-  m_UnitNickname: 
+  m_UnitNickname: Player 1
   m_Faction: Allies
-  m_MaxMana: 5
-  m_MaxDefense: 3
-  m_MaxHealth: 10
-  m_Experience: 0
-  m_Speed: 5
-  m_Nameplate: {fileID: 11465284, guid: 5846026beed07984eba03daa6cd87c58, type: 2}
+  m_BaseHealth: 5
+  m_HealthGrowth: 2.5
+  m_BaseMana: 2
+  m_ManaGrowth: 1
+  m_BaseDefense: 2
+  m_DefenseGrowth: 1
+  m_BaseLevel: 1
+  m_BaseSpeed: 5
+  m_SpeedGrowth: 0.2
   m_Parent: {fileID: 0}
-  m_CurrentRotation: {x: 0, y: 0, z: 0}
-  m_OriginalRotation: {x: 0, y: 0, z: 0}
 --- !u!135 &13513312
 SphereCollider:
   m_ObjectHideFlags: 1
